requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pandas==2.2.2
numpy==1.26.4
yfinance==0.2.43
ta==0.11.0
feedparser==6.0.11
nltk==3.9.1
pydantic==2.9.2
python-dotenv==1.0.1
alpaca-trade-api==3.2.0
SQLAlchemy==2.0.35
Jinja2==3.1.4
